requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-docx==0.8.11
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2

# Flask and related packages
flask-cors==4.0.0

